Move samples
